requirements.txt
arrow==0.10.0
boto==2.47.0
certifi==2017.4.17
chardet==3.0.4
Django==1.11.2
django-forms-bootstrap==3.1.0
django-gravatar2==1.4.0
djangorestframework==3.6.3
gravatar==0.1
idna==2.5
olefile==0.44
Pillow==4.1.1
python-dateutil==2.6.0
pytz==2017.2
requests==2.18.1
six==1.10.0
stripe==1.61.0
urllib3==1.21.1
